Moved notification mechanism into gRPC interface
