+++ nucliadb_node/src/bin/writer.rs
@@ -33,8 +33,8 @@ use nucliadb_core::tracing::*;
 use nucliadb_node::env;
 use nucliadb_node::reader::NodeReaderService;
 use nucliadb_node::telemetry::init_telemetry;
-use nucliadb_node::writer::grpc_driver::NodeWriterGRPCDriver;
-use nucliadb_node::writer::{NodeWriterEvent, NodeWriterMetadata, NodeWriterService};
+use nucliadb_node::writer::grpc_driver::{NodeWriterGRPCDriver, NodeWriterEvent, NodeWriterMetadata};
+use nucliadb_node::writer::NodeWriterService;
 use tokio::sync::mpsc::UnboundedReceiver;
 use tokio_stream::StreamExt;
 use tonic::transport::Server;
@@ -50,16 +50,16 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
 
     let start_bootstrap = Instant::now();
 
-    let (sender, receiver) = tokio::sync::mpsc::unbounded_channel();
     let node_metadata = get_node_metadata();
-    let mut node_writer_service = NodeWriterService::with_sender(sender);
+    let mut node_writer_service = NodeWriterService::new();
 
     std::fs::create_dir_all(env::shards_path())?;
     if !env::lazy_loading() {
         node_writer_service.load_shards()?;
     }
 
-    let grpc_driver = NodeWriterGRPCDriver::from(node_writer_service);
+    let (sender, receiver) = tokio::sync::mpsc::unbounded_channel();
+    let grpc_driver = NodeWriterGRPCDriver::from(node_writer_service).with_sender(sender);
     let host_key_path = env::host_key_path();
     let public_ip = env::public_ip().await;
     let chitchat_port = env::chitchat_port();
